Initial pass as lazy loading.
There are problems with this implementation that make it unsuitable for any use.
The initial request to a blob-like column will go back to the database for the value,
but so will every other request. There needs to be a way to have the class remember
that's already happened. This also applies to calling save(); we don't want existing
data replaced by an empty string except when explicitly set. So the current is broken,
but it's a start.